backend/requirements.txt
flask
flask-cors
supabase
python-dotenv
requests
google-generativeai
tqdm
gunicorn
httpx
httpcore
apscheduler
pytz